+++ .eslintrc.js
@@ -36,7 +36,18 @@ module.exports = {
       },
     },
     {
-      files: ["**/api/**/*.js", "**/api/**/*.ts"],
+      files: [
+        "**/api/**/*.js",
+        "**/api/**/*.ts",
+        "**/medusa-js/**/resources/**/*.ts",
+      ],
+      rules: {
+        "valid-jsdoc": ["off"],
+      },
+    },
+    {
+      // Medusa JS client
+      files: ["**/medusa-js/**/resources/**/*.ts"],
       rules: {
         "valid-jsdoc": ["off"],
       },