+++ packages/medusa/src/api/routes/admin/orders/index.ts
@@ -1,8 +1,8 @@
 import { Router } from "express"
 import { Order } from "../../../.."
 import middlewares from "../../../middlewares"
-import "reflect-metadata"
 import { DeleteResponse, PaginatedResponse } from "../../../../types/common"
+import "reflect-metadata"
 
 const route = Router()
 
@@ -205,7 +205,7 @@ export default (app) => {
   )
 
   /**
-   * Creates claim fulfillment
+   * Creates claim shipment
    */
   route.post(
     "/:id/claims/:claim_id/shipments",