+++ packages/medusa/src/api/routes/admin/regions/index.ts
@@ -110,3 +110,11 @@ export class FulfillmentOption {
 export class AdminGetRegionsRegionFulfillmentOptionsRes {
   fulfillment_options: FulfillmentOption[]
 }
+
+export * from "./list-regions"
+export * from "./update-region"
+export * from "./create-region"
+export * from "./add-country"
+export * from "./add-payment-provider"
+export * from "./add-fulfillment-provider"
+export * from "./set-metadata"