+++ packages/medusa/src/types/invites.ts
@@ -1,0 +1,5 @@
+import { Invite } from "../models/invite"
+
+export type ListInvite = Omit<Invite, "beforeInsert"> & {
+  token: string
+}